Layout Changes - Initiallizing a final layout. - Reorganized folders - Set up system for color constants - Creating default widgets
