+++ .vscode/launch.json
@@ -1,0 +1,28 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "onu-smart-system",
+            "request": "launch",
+            "type": "dart",
+            "program": "lib/pages/main.dart"
+        },
+        {
+            "name": "onu-smart-system (profile mode)",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "profile",
+            "program": "lib/pages/main.dart"
+        },
+        {
+            "name": "onu-smart-system (release mode)",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "release",
+            "program": "lib/pages/main.dart"
+        }
+    ]
+}
