+++ lib/constants.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+const onuOrange = Color.fromARGB(255, 255, 145, 0);
