--- lib/main.dart
@@ -1,357 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:onu_smart/pages/Options_Home.dart';
-import 'package:onu_smart/pages/Printer_home.dart';
-import 'package:onu_smart/pages/Tour_home.dart';
-import 'package:onu_smart/pages/fountain_home.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'firebase_options.dart';
-
-
-// void main() {
-//   runApp(const InspirationApp());
-// }
-
-Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
-  runApp(const InspirationApp());
-}
-
-
-class InspirationApp extends StatelessWidget {
-  const InspirationApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: const LoginPage(),
-      theme: ThemeData(
-        primaryColor: Colors.orange,
-        scaffoldBackgroundColor: Colors.black,
-      ),
-    );
-  }
-}
-
-class LoginPage extends StatelessWidget {
-  const LoginPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      
-      appBar: AppBar(
-        title: const Text('Launch'),
-        
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            ElevatedButton(
-              onPressed: () {
-                // Perform login logic here
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => const MainPage(),
-                  ),
-                );
-              },
-              style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.orange, // Set the background color to orange
-              foregroundColor: Colors.black, // Set the font color to black
-            ),
-              child: const Text('Login'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class MainPage extends StatelessWidget {
-  const MainPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('SMART JLK'),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            FountainLargeButton('Fountain Control'),
-            PrinterLargeButton('3D Print'),
-            TourLargeButton('Smart Tour'),
-            OptionsLargeButton('Options'),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class FountainLargeButton extends StatelessWidget {
-  final String text;
-
-  FountainLargeButton(this.text, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: ElevatedButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => const FountainHome(),
-            ),
-          );
-        },
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all(Colors.orange),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Text(
-            text,
-            style: const TextStyle(
-              color: Colors.black,
-              fontSize: 18.0,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-
-class PrinterLargeButton extends StatelessWidget {
-  final String text;
-
-  PrinterLargeButton(this.text, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: ElevatedButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => const PrinterHome(),
-            ),
-          );
-        },
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all(Colors.orange),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Text(
-            text,
-            style: const TextStyle(
-              color: Colors.black,
-              fontSize: 18.0,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-
-class TourLargeButton extends StatelessWidget {
-  final String text;
-
-  TourLargeButton(this.text, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: ElevatedButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => const TourHome(),
-            ),
-          );
-        },
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all(Colors.orange),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Text(
-            text,
-            style: const TextStyle(
-              color: Colors.black,
-              fontSize: 18.0,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class OptionsLargeButton extends StatelessWidget {
-  final String text;
-
-  OptionsLargeButton(this.text, {super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: ElevatedButton(
-        onPressed: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => const OptionsHome(),
-            ),
-          );
-        },
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all(Colors.orange),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Text(
-            text,
-            style: const TextStyle(
-              color: Colors.black,
-              fontSize: 18.0,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// import 'dart:async';
-// import 'dart:convert';
-
-// import 'package:flutter/material.dart';
-// import 'package:http/http.dart' as http;
-
-// Future<Album> createAlbum(String title) async {
-//   final response = await http.post(
-//     Uri.parse('https://jsonplaceholder.typicode.com/albums'),
-//     headers: <String, String>{
-//       'Content-Type': 'application/json; charset=UTF-8',
-//     },
-//     body: jsonEncode(<String, String>{
-//       'title': title,
-//     }),
-//   );
-
-//   if (response.statusCode == 201) {
-//     // If the server did return a 201 CREATED response,
-//     // then parse the JSON.
-//     return Album.fromJson(jsonDecode(response.body) as Map<String, dynamic>);
-//   } else {
-//     // If the server did not return a 201 CREATED response,
-//     // then throw an exception.
-//     throw Exception('Failed to create album.');
-//   }
-// }
-
-// class Album {
-//   final int id;
-//   final String title;
-
-//   const Album({required this.id, required this.title});
-
-//   factory Album.fromJson(Map<String, dynamic> json) {
-//     return Album(
-//       id: json['id'] as int,
-//       title: json['title'] as String,
-//     );
-//   }
-// }
-
-// void main() {
-//   runApp(const MyApp());
-// }
-
-// class MyApp extends StatefulWidget {
-//   const MyApp({super.key});
-
-//   @override
-//   State<MyApp> createState() {
-//     return _MyAppState();
-//   }
-// }
-
-// class _MyAppState extends State<MyApp> {
-//   final TextEditingController _controller = TextEditingController();
-//   Future<Album>? _futureAlbum;
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return MaterialApp(
-//       title: 'Create Data Example',
-//       theme: ThemeData(
-//         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-//       ),
-//       home: Scaffold(
-//         appBar: AppBar(
-//           title: const Text('Create Data Example'),
-//         ),
-//         body: Container(
-//           alignment: Alignment.center,
-//           padding: const EdgeInsets.all(8),
-//           child: (_futureAlbum == null) ? buildColumn() : buildFutureBuilder(),
-//         ),
-//       ),
-//     );
-//   }
-
-//   Column buildColumn() {
-//     return Column(
-//       mainAxisAlignment: MainAxisAlignment.center,
-//       children: <Widget>[
-//         TextField(
-//           controller: _controller,
-//           decoration: const InputDecoration(hintText: 'Enter Title'),
-//         ),
-//         ElevatedButton(
-//           onPressed: () {
-//             setState(() {
-//               _futureAlbum = createAlbum(_controller.text);
-//             });
-//           },
-//           child: const Text('Create Data'),
-//         ),
-//       ],
-//     );
-//   }
-
-//   FutureBuilder<Album> buildFutureBuilder() {
-//     return FutureBuilder<Album>(
-//       future: _futureAlbum,
-//       builder: (context, snapshot) {
-//         if (snapshot.hasData) {
-//           return Text(snapshot.data!.title);
-//         } else if (snapshot.hasError) {
-//           return Text('${snapshot.error}');
-//         }
-
-//         return const CircularProgressIndicator();
-//       },
-//     );
-//   }
-// }
